refactor(devices): move registration logic to separate file
